refactor: move dev stuff output webpack.js
